+++ Lilliputian/Game/CountdownTimer.cpp
@@ -1,13 +1,13 @@
 #include "CountdownTimer.hpp"
 
-void Lilliputian::CountdownTimer::start(double time)
+void Lilliputian::CountdownTimer::start(float time)
 {
 	this->startTime = time;
 	this->currentTime = startTime;
 	this->isActive = true;
 }
 
-void Lilliputian::CountdownTimer::update(double deltaTime)
+void Lilliputian::CountdownTimer::update(float deltaTime)
 {
 	if (this->isActive && this->currentTime > 0.0)
 		this->currentTime -= deltaTime;
