+++ Lilliputian/Game/PropertyAnimation.hpp
@@ -13,14 +13,14 @@ namespace Lilliputian
 	public:
 		EntityID entityID = NO_ENTITY;
 
-		double* targetDoubleProperty = NULL;
-		Vector<double> keyFrames;
-		double duration_s = 0.0;
-		double playbackSpeed_pct = 1.0;
-		double delay_s = 0.0;
-		double progress_s = 0.0;
-		double maximumValue = 0.0;
-		double minimumValue = 0.0;
+		float* targetDoubleProperty = NULL;
+		Vector<float> keyFrames;
+		float duration_s = 0.0;
+		float playbackSpeed_pct = 1.0;
+		float delay_s = 0.0;
+		float progress_s = 0.0;
+		float maximumValue = 0.0;
+		float minimumValue = 0.0;
 		uint8_t repeats = 0;
 		uint8_t repeatCount = 0;
 		bool isPlaying = false;
